requirements.txt
fastapi
uvicorn
pymysql
cryptography
python-dotenv
pydantic_settings

jose
PyJwt<2.10
passlib

google
google_auth_oauthlib

python-multipart